requirements.txt
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.2.0
PyYAML==5.1.1
python-coveralls
coveralls
cython
pycocotools
matplotlib
csv